Fixing off-by-one error in ASN.1 size calcuations
In each place that a byte-array index is used to calculate the size
of an ASN.1 object AND the index was being decremented as each byte
is written to the byte-array, an off-by-one error was causing the
length field to be one larger than was correct. The issue was fixed
by ceasing to use the decrement operator whenever calculating size
and decrementing by the appropriate amount after the size bytes
have been written to the byte array.

I have not researched more deeply, but I suspect that this is a
compiler-related optimization bug: that older versions of GCC
performed the decrement operation after the entire line of code
had executed but that at some point this changed and the
decrement occurs at some earlier point. Whether this is true
or not, it is probably a bad idea to be changing an index at
the same time as it is used for a length calculation. Also,
anywhere the lsb()/msb() pair of functions were used in that
order, the msb() function was receiving a decremented value,
and the bug did not manifest because the least significant
bits were thrown away.